+++ CHANGELOG.md
@@ -6,6 +6,7 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
 * Added
   * PnP editor: after manual selection, the component is appliet to all items where Comment and Footprint matches the current item
+  * PnP editor: item details order changed to: <index> | <footprint> | <comment>
 * Changed
   * DB components editor redesigned - now it's able to handle even 10'000 elements
 * Deprecated